style: clippy 清零（阶段3）
clippy --fix 自动修约20处 + 手动加5个类型别名(PduDecodeFn/ReadyCallbackFn/UnhealthyCallback) + run_connection 加 allow(too_many_arguments)。cargo clippy --workspace --lib 0 警告。验证：workspace lib + 四协议集成 + dynamic 测试全绿。

+++ crates/rsms-connector/src/transaction/smpp.rs
@@ -28,14 +28,14 @@ impl SmppSubmit {
                     .as_bytes()
                     .first()
                     .copied()
-                    .unwrap_or(b'0') as u8,
+                    .unwrap_or(b'0'),
                 self.inner
                     .destination_addr
                     .as_bytes()
                     .first()
                     .copied()
-                    .unwrap_or(b'0') as u8,
-                self.inner.short_message.first().copied().unwrap_or(b'0') as u8,
+                    .unwrap_or(b'0'),
+                self.inner.short_message.first().copied().unwrap_or(b'0'),
                 self.inner.short_message.len().min(5) as u8,
                 0,
                 0,
